+++ tests-integration/test_constructor.py
@@ -22,6 +22,21 @@ class TestConstructor(unittest.TestCase):
         with self.assertRaises(botocore.exceptions.NoRegionError):
             apply('tests-integration/fixtures/awstaga.yaml', False)
 
+    def test_constructor_with_config_not_having_resource_tags(self):
+        # should raise no region error
+        with self.assertRaises(botocore.exceptions.NoRegionError):
+            apply('tests-integration/fixtures/awstaga-no-resource-tags.yaml', False)
+
+    def test_constructor_with_config_not_having_resource_tagsetnames(self):
+        # should raise no region error
+        with self.assertRaises(botocore.exceptions.NoRegionError):
+            apply('tests-integration/fixtures/awstaga-no-resource-tagsetnames.yaml', False)
+
+    def test_constructor_with_config_not_having_tagsets(self):
+        # should raise no region error
+        with self.assertRaises(botocore.exceptions.NoRegionError):
+            apply('tests-integration/fixtures/awstaga-no-tagsets.yaml', False)
+
     def test_constructor_with_aws_region_without_keys(self):
         os.environ['AWS_DEFAULT_REGION'] = 'ap-southeast-2'
         # should raise client error