+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add dry-run support
 - Add Python 3.12 support
 
+### Changed
+- Allow tagsets, resource tags, and resource tagsetnames to be optional #1
+
 ### Removed
 - Remove unused certifi dependency